Refactor: Migrate API types to domain-driven modular structure
