+++ web-app/src/domain/types/index.ts
@@ -1,0 +1,21 @@
+// === ENTITIES ===
+export * from './entities/user.types'
+export * from './entities/group.types'
+export * from './entities/task.types'
+export * from './entities/action.types'
+export * from './entities/tag.types'
+
+// === PAYLOADS ===
+export * from './payloads/auth.payloads'
+export * from './payloads/group.payloads'
+export * from './payloads/task.payloads'
+export * from './payloads/action.payloads'
+export * from './payloads/tag.payloads'
+
+// === RESPONSES ===
+export * from './responses/auth.responses'
+export * from './responses/group.responses'
+export * from './responses/task.responses'
+export * from './responses/action.responses'
+export * from './responses/overview.responses'
+export * from './responses/common.responses'
