+++ web-app/src/domain/types/payloads/action.payloads.ts
@@ -1,0 +1,4 @@
+export interface CreateActionPayload {
+  taskId: number
+  date: string
+}
