+++ web-app/src/domain/types/payloads/auth.payloads.ts
@@ -1,0 +1,13 @@
+export interface LoginPayload {
+  email: string
+  password: string
+}
+
+export interface RegisterPayload {
+  nom: string
+  prenom: string
+  pseudo: string
+  email: string
+  password: string
+  icone?: string
+}
