+++ web-app/src/domain/types/payloads/group.payloads.ts
@@ -1,0 +1,8 @@
+export interface CreateGroupPayload {
+  nom: string
+}
+
+export interface JoinGroupPayload {
+  groupId: number
+  code: string
+}
