+++ web-app/src/domain/types/payloads/tag.payloads.ts
@@ -1,0 +1,5 @@
+export interface CreateTagPayload {
+  label: string
+  color: string
+  groupId: number
+}
