+++ web-app/src/domain/types/responses/action.responses.ts
@@ -1,0 +1,13 @@
+import type { Action } from '../entities/action.types'
+
+export interface CreateActionResponse {
+  message: string
+  action: Action
+  totalDone: number
+}
+
+export interface GetRecentActionsResponse {
+  message: string
+  actions: Action[]
+  total: number
+}
