+++ web-app/src/domain/types/responses/auth.responses.ts
@@ -1,0 +1,11 @@
+import type { User } from '../entities/user.types'
+
+export interface AuthResponse {
+  token: string
+  user: User
+}
+
+export interface getProfileResponse {
+  message: string
+  user: User
+}
