+++ web-app/src/domain/types/responses/common.responses.ts
@@ -1,0 +1,7 @@
+// Types d'erreur
+export interface ApiError {
+  message: string
+  statusCode: number
+  timestamp: string
+  path: string
+}
