+++ web-app/src/domain/types/responses/task.responses.ts
@@ -1,0 +1,11 @@
+import type { Task } from '../entities/task.types'
+
+export interface CreateTaskResponse {
+  message: string
+  task: Task
+}
+
+export interface UpdateTaskResponse {
+  message: string
+  task: Task
+}
