+++ web-app/src/presentation/components/molecules/RegisterForm.vue
@@ -113,7 +113,7 @@ import {
 } from '@heroicons/vue/24/outline'
 import BaseInput from '@/presentation/components/atoms/BaseInput.vue'
 import BaseButton from '@/presentation/components/atoms/BaseButton.vue'
-import type { RegisterPayload } from '@/shared/types/api'
+import type { RegisterPayload } from '@/domain/types'
 
 interface Props {
   loading?: boolean
