+++ web-app/src/presentation/components/molecules/TagStatsCard.vue
@@ -33,7 +33,7 @@
 
 <script setup lang="ts">
 import { computed } from 'vue'
-import type { Action } from '@/shared/types/api'
+import type { Action } from '@/domain/types'
 
 interface Props {
   tagStats: Record<number, number>
