+++ web-app/src/presentation/components/molecules/TaskList.vue
@@ -34,7 +34,7 @@
 
 <script setup lang="ts">
 import { computed } from 'vue'
-import type { Task, Tag } from '@/shared/types/api'
+import type { Task, Tag } from '@/domain/types'
 import TaskCard from '@/presentation/components/atoms/TaskCard.vue'
 
 interface Props {
