+++ web-app/src/presentation/views/GroupDetailView.vue
@@ -181,7 +181,7 @@ import { ref, computed, onMounted, onUnmounted } from 'vue'
 import { useRoute, useRouter } from 'vue-router'
 import { useGroupStore } from '@/domain/stores/groupStore'
 import { useTasksStore } from '@/domain/stores/tasksStore'
-import type { Tag, Task, CreateTaskPayload, CreateTagPayload, UpdateTaskPayload } from '@/shared/types/api'
+import type { Tag, Task, CreateTaskPayload, CreateTagPayload, UpdateTaskPayload } from '@/domain/types'
 import AppLayout from '@/presentation/layouts/AppLayout.vue'
 import TagFilter from '@/presentation/components/molecules/TagFilter.vue'
 import TaskList from '@/presentation/components/molecules/TaskList.vue'
