--- web-app/src/shared/types/api.ts
@@ -1,301 +1,0 @@
-// Entités API
-export interface User {
-  id: number
-  nom: string
-  prenom: string
-  pseudo: string
-  icone?: string
-  email: string
-  createdAt: string
-  updatedAt: string
-}
-
-export interface UserWithActions extends User {
-  actions: Action[]
-}
-
-export interface Group {
-  id: number
-  nom: string
-  code: string
-  users: User[]
-  tasks: Task[]
-  actions: Action[]
-  tags: Tag[]
-  createdAt: string
-  updatedAt: string
-}
-
-export enum UniteFrequence {
-  JOUR = 'jour',
-  SEMAINE = 'semaine',
-  MOIS = 'mois'
-}
-
-export interface Task {
-  id: number
-  label: string
-  iconUrl?: string
-  frequenceEstimee: number
-  uniteFrequence: UniteFrequence
-  points: number
-  group: Group
-  tag?: Tag
-  actions: Action[]
-  userTaskState?: UserTaskState | null
-  createdAt: string
-  updatedAt: string
-}
-
-export interface CreateTaskPayload {
-  label: string
-  iconUrl?: string
-  frequenceEstimee: number
-  uniteFrequence: UniteFrequence
-  groupId: number
-  tagId?: number
-  points: number
-}
-
-export interface UpdateTaskPayload {
-  label: string
-  iconUrl?: string
-  frequenceEstimee: number
-  uniteFrequence: UniteFrequence
-  tagId?: number
-  points: number
-}
-
-export interface CreateTaskResponse {
-  message: string
-  task: Task
-}
-
-export interface UpdateTaskResponse {
-  message: string
-  task: Task
-}
-
-export interface Action {
-  id: number
-  date: string
-  isHelpingHand: boolean
-  task: Task
-  user: User
-  group: Group
-  createdAt: string
-  updatedAt: string
-}
-
-export interface Tag {
-  id: number
-  label: string
-  color: string
-  group: Group
-  tasks: Task[]
-  createdAt: string
-  updatedAt: string
-}
-
-export interface UserTaskState {
-  id: number
-  taskId: number
-  userId: number
-  isAcknowledged: boolean
-  isConcerned: boolean
-  acknowledgedAt?: string
-  concernedAt?: string
-  createdAt: string
-  updatedAt: string
-}
-
-// Payloads pour les requêtes
-export interface LoginPayload {
-  email: string
-  password: string
-}
-
-export interface RegisterPayload {
-  nom: string
-  prenom: string
-  pseudo: string
-  email: string
-  password: string
-  icone?: string
-}
-
-export interface FetchGroupResponse {
-  message: string
-  group: Group
-}
-
-export interface CreateGroupPayload {
-  nom: string
-}
-
-export interface CreateGroupResponse {
-  message: string
-  group: Group
-}
-
-export interface JoinGroupPayload {
-  groupId: number
-  code: string
-}
-
-export interface CreateTagPayload {
-  label: string
-  color: string
-  groupId: number
-}
-
-export interface CreateActionPayload {
-  taskId: number
-  date: string
-}
-
-export interface CreateActionResponse {
-  message: string
-  action: Action
-  totalDone: number
-}
-
-export interface GetRecentActionsResponse {
-  message: string
-  actions: Action[]
-  total: number
-}
-
-export interface GroupStatistics {
-  totalActions: number
-  actionsByUser: Record<number, number>
-  actionsByTask: Record<number, number>
-  actionsByTag: Record<number, number>
-  totalWeight: number
-  actions: Action[]
-}
-
-// Types pour la réponse Overview
-export interface MonthlyVolumeItem {
-  taskId: number
-  taskLabel: string
-  monthlyFrequency: number
-  points: number
-  monthlyPoints: number
-  tag: {
-    id: number
-    label: string
-    color: string
-  } | null
-}
-
-export interface ActionItem {
-  id: number
-  taskLabel: string
-  points: number
-  isHelpingHand: boolean
-  tag: {
-    id: number
-    label: string
-    color: string
-  } | null
-}
-
-export interface UserDayActions {
-  userName: string
-  actions: ActionItem[]
-}
-
-export interface ActionsByDayAndUser {
-  [dayKey: string]: {
-    [userId: number]: UserDayActions
-  }
-}
-
-export interface HelpingHandItem {
-  id: number
-  date: string
-  taskLabel: string
-  points: number
-  tag: {
-    id: number
-    label: string
-    color: string
-  } | null
-}
-
-export interface HelpingHandByUser {
-  [userId: number]: {
-    userName: string
-    helpingHands: HelpingHandItem[]
-  }
-}
-
-export interface ActionsByCategory {
-  tagId: number | null
-  tagLabel: string
-  tagColor: string
-  totalMonthlyVolume: number
-  completedThisMonth: number
-  completionPercentage: number
-  actionsCount: number
-  tasksInCategory: number
-}
-
-export interface Overview {
-  totalTasksVolume: number,
-  totalDone: number,
-  actions: Action[],
-  users: UserWithActions[],
-  tasks: Task[],
-}
-
-export interface PersonalGoal {
-  user: User
-  doneThisMonth: number
-}
-
-export interface GetOverviewResponse {
-  message: string
-  overview: Overview
-}
-
-export interface UpdateUserTaskStatePayload {
-  isAcknowledged?: boolean
-  isConcerned?: boolean
-}
-
-// Responses de l'API
-export interface AuthResponse {
-  token: string
-  user: User
-}
-
-export interface GroupSearchResponse {
-  message: string
-  groups: Group[]
-}
-
-export interface getProfileResponse {
-  message: string
-  user: User
-}
-
-export interface UserGroupResponse {
-  message: string
-  groups: Group[]
-}
-
-export interface StatisticsResponse {
-  totalActions: number
-  actionsByUser: Record<string, number>
-  actionsByTask: Record<string, number>
-}
-
-// Types d'erreur
-export interface ApiError {
-  message: string
-  statusCode: number
-  timestamp: string
-  path: string
-}
